Added integration tests for basic functionality
Added tests for simple C, C++ and Rust programs doing allocations.
The included tests will build the program, run it with allocscope-trace,
generate a text report with allocscope-view, parse the report and verify
that the report contents match expected values.

This commit also adds a test using realloc(), a test which traces a
program doing alloctions from several threads, and a test case which
will attach the trace by PID to an already running process.

The script 'test.sh' in the root of the repo can be used to run all
integration tests.